requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
numpy
torch
transformers
scikit-learn
PyPDF2
python-docx
requests